refactor(reth-bench): dedupe read_input and load_jwt_secret helpers
